src/auto_multi_containersV0/requirements.txt
numpy==2.0.0
requests
PyYAML